requirements.txt
streamlit==1.28.1
pandas==2.1.1
gspread==5.10.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-auth==2.25.2
gspread-dataframe==3.3.1
openpyxl==3.11.0
